+++ hobj/benchmarks/leedicarlo_benchmarks/MutatorHighVar/mut_highvar_benchmark.py
@@ -1,7 +1,7 @@
 import numpy as np
 import xarray as xr
 
-from hobj.data.behavioral_data import human_data as human_data
+from hobj.data.behavior import human_data as human_data
 from hobj.learning_models import learning_model as lm
 from hobj.utils import stats as stats
